fix(topbar): remove circle dependency
