--- packages/admin-ui/src/components/Topbar/index.ts
@@ -1,1 +1,0 @@
-export * from './topbar'
