+++ packages/admin-ui/src/components/Topbar/index.tsx
@@ -1,8 +1,10 @@
 import type { ComponentPropsWithRef } from 'react'
-import React from 'react'
+import React, { Fragment } from 'react'
 import { jsx } from '@vtex/admin-ui-react'
 import { experimental_Bleed as Bleed } from '../../experimental'
-import { TopbarSkeleton } from './topbar-skeleton'
+import { Set } from '../Set'
+import { Center } from '../Center'
+import { Skeleton } from '../Skeleton'
 
 /**
  * Topbar component.
@@ -104,6 +106,43 @@ export const TopbarEnd = jsx('div')(
     },
   }
 )
+
+/**
+ * Topbar loading Skeleton
+ */
+export function TopbarSkeleton() {
+  return (
+    <Fragment>
+      <TopbarStart>
+        <Set spacing="$l">
+          <Center csx={{ padding: '$xs' }}>
+            <Skeleton csx={{ size: '1.5rem' }} />
+          </Center>
+          <Skeleton csx={{ width: '8rem', height: '1rem' }} />
+        </Set>
+      </TopbarStart>
+      <TopbarCenter>
+        <Skeleton csx={{ width: '13.75rem', height: '2rem' }} />
+      </TopbarCenter>
+      <TopbarEnd>
+        <Set spacing="$m">
+          <Center csx={{ padding: '$xs' }}>
+            <Skeleton csx={{ height: '1.25rem', width: '5.563rem' }} />
+          </Center>
+          <Center csx={{ padding: '$xs' }}>
+            <Skeleton csx={{ size: '1.25rem' }} />
+          </Center>
+          <Center csx={{ padding: '$xs' }}>
+            <Skeleton csx={{ size: '1.25rem' }} />
+          </Center>
+          <Center csx={{ padding: '$xs' }}>
+            <Skeleton csx={{ size: '1.5rem' }} />
+          </Center>
+        </Set>
+      </TopbarEnd>
+    </Fragment>
+  )
+}
 
 export interface TopbarOptions {
   /**
